refactor(cymbal): remove Catalog/fingerprinting back-edges from frames
Move the RawFrame `Resolve<Catalog>` dispatcher into
symbolication::resolve (alongside the per-lang impls), and move
`impl FingerprintComponent for Frame` into the processing fingerprinting
module. The frames module no longer depends on the symbol-store Catalog or
on fingerprinting — it's now pure frame data, the prerequisite for
relocating it into core/types.

No behavior change; cargo check + clippy green; langs/resolve/fingerprinting
tests pass.
