+++ rust/cymbal/src/core/frames/mod.rs
@@ -6,18 +6,14 @@ use serde::{Deserialize, Serialize};
 use serde_json::Value;
 
 use crate::{
-    error::UnhandledError,
-    fingerprinting::{FingerprintBuilder, FingerprintComponent, FingerprintRecordPart},
     langs::{
         apple::RawAppleFrame, custom::CustomFrame, dart::RawDartFrame, go::RawGoFrame,
         hermes::RawHermesFrame, java::RawJavaFrame, js::RawJSFrame, native::DebugImage,
         node::RawNodeFrame, php::RawPHPFrame, python::RawPythonFrame, ruby::RawRubyFrame,
         rust::RawRustFrame,
     },
-    metric_consts::{FRAME_NOT_RESOLVED, FRAME_RESOLVED, LEGACY_JS_FRAME_RESOLVED, PER_FRAME_TIME},
+    metric_consts::FRAME_NOT_RESOLVED,
     sanitize_source_line,
-    symbolication::resolve::Resolve,
-    symbolication::symbol_store::Catalog,
 };
 
 /// Records the metric and tracing line for a single failed-frame construction. Each
@@ -75,79 +71,6 @@ pub enum RawFrame {
     // TODO - remove once we're happy no clients are using this anymore
     #[serde(rename = "javascript")]
     LegacyJS(RawJSFrame),
-}
-
-impl Resolve<Catalog> for RawFrame {
-    async fn resolve(
-        &self,
-        team_id: i32,
-        catalog: &Catalog,
-        debug_images: &[DebugImage],
-    ) -> Result<Vec<Frame>, UnhandledError> {
-        let frame_resolve_time = common_metrics::timing_guard(PER_FRAME_TIME, &[]);
-        // Catalog-backed frames resolve via `Resolve`; catalog-free frames convert
-        // directly with `From<&RawX> for Frame` (no symbol resolution needed).
-        let (res, lang_tag): (Result<Vec<Frame>, UnhandledError>, &str) = match self {
-            RawFrame::JavaScriptWeb(frame) => {
-                (frame.resolve(team_id, catalog, debug_images).await, "javascript")
-            }
-            RawFrame::LegacyJS(frame) => {
-                // TODO: monitor this metric and remove the legacy frame type when it hits 0
-                metrics::counter!(LEGACY_JS_FRAME_RESOLVED).increment(1);
-                (frame.resolve(team_id, catalog, debug_images).await, "javascript")
-            }
-            RawFrame::JavaScriptNode(frame) => {
-                (frame.resolve(team_id, catalog, debug_images).await, "javascript")
-            }
-
-            RawFrame::Dart(frame) => (Ok(vec![frame.into()]), "dart"),
-            RawFrame::Apple(frame) => {
-                (frame.resolve(team_id, catalog, debug_images).await, "apple")
-            }
-            RawFrame::Php(frame) => (Ok(vec![frame.into()]), "php"),
-            RawFrame::Python(frame) => (Ok(vec![frame.into()]), "python"),
-            RawFrame::Ruby(frame) => (Ok(vec![frame.into()]), "ruby"),
-            RawFrame::Rust(frame) => (Ok(vec![frame.into()]), "rust"),
-            RawFrame::Custom(frame) => (Ok(vec![frame.into()]), "custom"),
-            RawFrame::Go(frame) => (Ok(vec![frame.into()]), "go"),
-            RawFrame::Hermes(frame) => {
-                (frame.resolve(team_id, catalog, debug_images).await, "hermes")
-            }
-            RawFrame::Java(frame) => (frame.resolve(team_id, catalog, debug_images).await, "java"),
-        };
-
-        // The raw id of the frame is set after it's resolved.
-        let res = res.map(|mut fs| {
-            fs.iter_mut()
-                .enumerate()
-                .for_each(|(index, f)| f.frame_id = self.frame_id(team_id, index, debug_images));
-            fs
-        });
-
-        if res.is_err() {
-            frame_resolve_time.label("outcome", "failed")
-        } else {
-            frame_resolve_time.label("outcome", "success")
-        }
-        .label("lang", lang_tag)
-        .fin();
-
-        if let Ok(frames) = &res {
-            for frame in frames {
-                if frame.resolved {
-                    metrics::counter!(FRAME_RESOLVED, "lang" => lang_tag).increment(1);
-                }
-                // Failure metrics are emitted by the language-specific `From` impls in
-                // `langs/*.rs` at the moment of frame construction, where the typed error
-                // is in scope (so we can call `metric_reason()` directly). This avoids
-                // having to carry the typed error on the `Frame` struct just to recover
-                // the metric label, which previously required a custom serializer plus
-                // `skip_deserializing` and silently dropped failure reasons on PG round-trip.
-            }
-        }
-
-        res
-    }
 }
 
 impl RawFrame {
@@ -255,55 +178,6 @@ pub struct Context {
 pub struct ContextLine {
     pub number: u32,
     pub line: String,
-}
-
-impl FingerprintComponent for Frame {
-    fn update(&self, fp: &mut FingerprintBuilder) {
-        let get_part = |s: &FrameId, p: Vec<&str>| FingerprintRecordPart::Frame {
-            raw_id: s.to_string(),
-            pieces: p.into_iter().map(String::from).collect(),
-        };
-
-        let mut included_pieces = Vec::new();
-
-        // Include source and module in the fingerprint either way
-        if let Some(source) = &self.source {
-            fp.update(source.as_bytes());
-            included_pieces.push("Source file name");
-        }
-
-        if let Some(module) = &self.module {
-            fp.update(module.as_bytes());
-            included_pieces.push("Module name");
-        }
-
-        // If we've resolved this frame, include function name, and then return
-        if let Some(resolved) = &self.resolved_name {
-            fp.update(resolved.as_bytes());
-            included_pieces.push("Resolved function name");
-
-            fp.add_part(get_part(&self.frame_id, included_pieces));
-            return;
-        }
-
-        // Otherwise, get more granular
-        fp.update(self.mangled_name.as_bytes());
-        included_pieces.push("Mangled function name");
-
-        if let Some(line) = self.line {
-            fp.update(line.to_string().as_bytes());
-            included_pieces.push("Line number");
-        }
-
-        if let Some(column) = self.column {
-            fp.update(column.to_string().as_bytes());
-            included_pieces.push("Column number");
-        }
-
-        fp.update(self.lang.as_bytes());
-        included_pieces.push("Language");
-        fp.add_part(get_part(&self.frame_id, included_pieces));
-    }
 }
 
 impl Frame {
